GH-269: The shared schema is now moved and everybody agrees.
