+++ masq_lib/src/constants.rs
@@ -3,3 +3,8 @@
 pub const HIGHEST_RANDOM_CLANDESTINE_PORT: u16 = 9999;
 pub const HTTP_PORT: u16 = 80;
 pub const TLS_PORT: u16 = 443;
+pub const DEFAULT_CHAIN_NAME: &str = "mainnet";
+pub const DEFAULT_GAS_PRICE: &str = "1";
+pub const LOWEST_USABLE_INSECURE_PORT: u16 = 1025;
+pub const HIGHEST_USABLE_PORT: u16 = 65535;
+pub const DEFAULT_UI_PORT: u16 = 5333;
