+++ node/src/bootstrapper.rs
@@ -22,7 +22,6 @@ use crate::server_initializer::LoggerInitializerWrapper;
 use crate::sub_lib::accountant;
 use crate::sub_lib::accountant::AccountantConfig;
 use crate::sub_lib::blockchain_bridge::BlockchainBridgeConfig;
-use crate::sub_lib::crash_point::CrashPoint;
 use crate::sub_lib::cryptde::CryptDE;
 use crate::sub_lib::cryptde_null::CryptDENull;
 use crate::sub_lib::cryptde_real::CryptDEReal;
@@ -37,7 +36,8 @@ use futures::try_ready;
 use itertools::Itertools;
 use log::LevelFilter;
 use masq_lib::command::StdStreams;
-use masq_lib::ui_gateway::DEFAULT_UI_PORT;
+use masq_lib::constants::DEFAULT_UI_PORT;
+use masq_lib::crash_point::CrashPoint;
 use std::collections::HashMap;
 use std::env::var;
 use std::fmt::{Debug, Error, Formatter};
