+++ node/src/crash_test_dummy.rs
@@ -1,8 +1,8 @@
 // Copyright (c) 2017-2019, Substratum LLC (https://substratum.net) and/or its affiliates. All rights reserved.
 
-use crate::sub_lib::crash_point::CrashPoint;
 use crate::sub_lib::logger::Logger;
 use futures::Async;
+use masq_lib::crash_point::CrashPoint;
 use tokio::prelude::future::Future;
 
 pub struct CrashTestDummy<C> {
